requirements.txt
fastapi
uvicorn[standard]
google-auth
google-auth-oauthlib
google-api-python-client
slack_sdk
twilio
python-dotenv
requests
transformers
torch
